Fix the UsageError in sereto repl
The error is fixed upstream, but there is no release yet. Reference
directly the git commit in the dependencies.

If you already have the click-repl 0.3.0 installed, you need to remove
it manually first, as the version from the specific git commit still has the
same version. As a result, it would not be replaced automatically.